requirements.txt
streamlit
python-dotenv

langchain
langchain-core
langchain-community
langchain-text-splitters

langchain-groq

sentence-transformers
langchain-huggingface

faiss-cpu

pypdf

google-search-results